+++ src/browser/ajax.js
@@ -1,13 +1,6 @@
-import atomic from 'atomic';
-const ajax = typeof window !== 'undefined' && atomic(window);
-
 export const loadViaAjax = (url, onLoaded) => {
-  ajax.get(url)
-    .success((data) => {
-      onLoaded(null, data);
-    })
-    .error((e, xhr) => {
-      onLoaded(e);
-    })
-  ;
+  window.fetch(url)
+    .then(response => response.json())
+    .then(data => onLoaded(null, data))
+    .catch(e => onLoaded(e));
 };