+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 
 # Version 1
 
-- [ ] use fetch instead of atomic
+- [x] use fetch instead of atomic
 - [ ] move to kavun as test runner, faster, simpler, smaller
 - [ ] remove workshop banner
 - [ ] classnames seems to be used rarely